+++ tests/str_test.cc
@@ -403,3 +403,31 @@ TEST(Str, append_n_chars) {
   ASSERT_STREQ(str.str, "     ");
   str_reset(&str);
 }
+
+TEST(Str, common_prefix) {
+  str_t str;
+
+  str_init(&str, 100);
+
+  str_set(&str, "abc.cde.123");
+  ASSERT_EQ(str_common_prefix(&str, "abc.123"), RET_OK);
+  ASSERT_STREQ(str.str, "abc.");
+  
+  str_set(&str, "abc.cde.123");
+  ASSERT_EQ(str_common_prefix(&str, "abc.c123"), RET_OK);
+  ASSERT_STREQ(str.str, "abc.c");
+  
+  str_set(&str, "abc.cde.123");
+  ASSERT_EQ(str_common_prefix(&str, "c123"), RET_OK);
+  ASSERT_STREQ(str.str, "");
+  
+  str_set(&str, "abc.cde.123");
+  ASSERT_EQ(str_common_prefix(&str, "abc.cde.123"), RET_OK);
+  ASSERT_STREQ(str.str, "abc.cde.123");
+  
+  str_set(&str, "");
+  ASSERT_EQ(str_common_prefix(&str, "abc.cde.123"), RET_OK);
+  ASSERT_STREQ(str.str, "");
+  
+  str_reset(&str);
+}